requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
pyperclip>=1.8.0
customtkinter>=5.2.0
pyinstaller>=6.0.0
tweepy>=4.14.0
instagrapi>=2.1.0
Pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
numpy>=1.24.0
